Extract window_wrapper and the update loop to their own modules
